+++ source/core/dem_properties.cc
@@ -19,10 +19,10 @@ namespace DEM
           PropertiesIndex::n_properties);
         properties[PropertiesIndex::type]    = std::make_pair("type", 1);
         properties[PropertiesIndex::dp]      = std::make_pair("diameter", 1);
-        properties[PropertiesIndex::v_x]     = std::make_pair("velocity", dim);
+        properties[PropertiesIndex::v_x]     = std::make_pair("velocity", 3);
         properties[PropertiesIndex::v_y]     = std::make_pair("velocity", 1);
         properties[PropertiesIndex::v_z]     = std::make_pair("velocity", 1);
-        properties[PropertiesIndex::omega_x] = std::make_pair("omega", dim);
+        properties[PropertiesIndex::omega_x] = std::make_pair("omega", 3);
         properties[PropertiesIndex::omega_y] = std::make_pair("omega", 1);
         properties[PropertiesIndex::omega_z] = std::make_pair("omega", 1);
         properties[PropertiesIndex::mass]    = std::make_pair("mass", 1);
@@ -37,20 +37,20 @@ namespace DEM
           PropertiesIndex::n_properties);
         properties[PropertiesIndex::type]    = std::make_pair("type", 1);
         properties[PropertiesIndex::dp]      = std::make_pair("diameter", 1);
-        properties[PropertiesIndex::v_x]     = std::make_pair("velocity", dim);
+        properties[PropertiesIndex::v_x]     = std::make_pair("velocity", 3);
         properties[PropertiesIndex::v_y]     = std::make_pair("velocity", 1);
         properties[PropertiesIndex::v_z]     = std::make_pair("velocity", 1);
-        properties[PropertiesIndex::omega_x] = std::make_pair("omega", dim);
+        properties[PropertiesIndex::omega_x] = std::make_pair("omega", 3);
         properties[PropertiesIndex::omega_y] = std::make_pair("omega", 1);
         properties[PropertiesIndex::omega_z] = std::make_pair("omega", 1);
         properties[PropertiesIndex::fem_force_x] =
-          std::make_pair("fem_force", dim);
+          std::make_pair("fem_force", 3);
         properties[PropertiesIndex::fem_force_y] =
           std::make_pair("fem_force", 1);
         properties[PropertiesIndex::fem_force_z] =
           std::make_pair("fem_force", 1);
         properties[PropertiesIndex::fem_torque_x] =
-          std::make_pair("fem_torque", dim);
+          std::make_pair("fem_torque", 3);
         properties[PropertiesIndex::fem_torque_y] =
           std::make_pair("fem_torque", 1);
         properties[PropertiesIndex::fem_torque_z] =